[HTM-752] restore FeaturesControllerPostgresIntegrationTest
pickup testcases from FeaturesControllerIntegrationTest and remove it
