+++ pom.xml
@@ -750,7 +750,6 @@ SPDX-License-Identifier: MIT
                         </path>
                     </annotationProcessorPaths>
                     <testExcludes>
-                        <exclude>**/FeaturesController*.java</exclude>
                         <exclude>**/LayerExportController*.java</exclude>
                     </testExcludes>
                 </configuration>
